Crate srml_balances

Balances Module

The Balances module provides functionality for handling accounts and balances.

balances::Trait

Call

Module

Overview

The Balances module provides functions for:

Getting and setting free balances.

Retrieving total, reserved and unreserved balances.

Repatriating a reserved balance to a beneficiary account that exists.

Transferring a balance between accounts (when not reserved).

Slashing an account balance.

Account creation and removal.

Managing total issuance.

Setting and managing locks.

Terminology

Existential Deposit: The minimum balance required to create or keep an account open. This prevents "dust accounts" from filling storage.

Total Issuance: The total number of units in existence in a system.

Reaping an account: The act of removing an account by resetting its nonce. Happens after its balance is set to zero.

Free Balance: The portion of a balance that is not reserved. The free balance is the only balance that matters for most operations. When this balance falls below the existential deposit, most functionality of the account is removed. When both it and the reserved balance are deleted, then the account is said to be dead.

Reserved Balance: Reserved balance still belongs to the account holder, but is suspended. Reserved balance can still be slashed, but only after all the free balance has been slashed. If the reserved balance falls below the existential deposit then it and any related functionality will be deleted. When both it and the free balance are deleted, then the account is said to be dead.

Imbalance: A condition when some funds were credited or debited without equal and opposite accounting (i.e. a difference between total issuance and account balances). Functions that result in an imbalance will return an object of the Imbalance trait that can be managed within your runtime logic. (If an imbalance is simply dropped, it should automatically maintain any book-keeping such as total issuance.)

Lock: A freeze on a specified amount of an account's free balance until a specified block number. Multiple locks always operate over the same funds, so they "overlay" rather than "stack".

Vesting: Similar to a lock, this is another, but independent, liquidity restriction that reduces linearly over time.

Implementations

The Balances module provides implementations for the following traits. If these traits provide the functionality that you need, then you can avoid coupling with the Balances module.

Currency: Functions for dealing with a fungible assets system.

ReservableCurrency: Functions for dealing with assets that can be reserved from an account.

LockableCurrency: Functions for dealing with accounts that allow liquidity restrictions.

Imbalance: Functions for handling imbalances between total issuance in the system and account balances. Must be used when a function creates new funds (e.g. a reward) or destroys some funds (e.g. a system fee).

MakePayment: Simple trait designed for hooking into a transaction payment.

IsDeadAccount: Determiner to say whether a given account is unused.

Interface

Dispatchable Functions

transfer - Transfer some liquid free balance to another account.

set_balance - Set the balances of a given account. The origin of this call must be root.

Public Functions

vesting_balance - Get the amount that is currently being vested and cannot be transferred out of this account.

Usage

The following examples show how to use the Balances module in your custom module.

Examples from the SRML

The Contract module uses the Currency trait to handle gas payment, and its types inherit from Currency:

use srml_support::traits::Currency;

pub type BalanceOf<T> = <<T as Trait>::Currency as Currency<<T as system::Trait>::AccountId>>::Balance;
pub type NegativeImbalanceOf<T> = <<T as Trait>::Currency as Currency<<T as system::Trait>::AccountId>>::NegativeImbalance;

The Staking module uses the LockableCurrency trait to lock a stash account's funds:

use srml_support::traits::{WithdrawReasons, LockableCurrency};
use sr_primitives::traits::Bounded;
pub trait Trait: system::Trait {
	type Currency: LockableCurrency<Self::AccountId, Moment=Self::BlockNumber>;
}

fn update_ledger<T: Trait>(
	controller: &T::AccountId,
	ledger: &StakingLedger<T>
) {
	T::Currency::set_lock(
		STAKING_ID,
		&ledger.stash,
		ledger.total,
		T::BlockNumber::max_value(),
		WithdrawReasons::all()
	);
	// <Ledger<T>>::insert(controller, ledger); // Commented out as we don't have access to Staking's storage here.
}

Genesis config

The Balances module depends on the GenesisConfig.

Assumptions

Total issued balanced of all accounts should be less than Trait::Balance::max_value().

Structs

BalanceLock

DefaultInstance
Default module instance

FreeBalance
The 'free' balance of a given account.

GenesisConfig

Instance0
Module instance

Instance1
Module instance

Instance2
Module instance

Instance3
Module instance

Instance4
Module instance

Instance5
Module instance

Instance6
Module instance

Instance7
Module instance

Instance8
Module instance

Instance9
Module instance

Instance10
Module instance

Instance11
Module instance

Instance12
Module instance

Instance13
Module instance

Instance14
Module instance

Instance15
Module instance

Locks
Any liquidity locks on some account balances.

Module

NegativeImbalance
Opaque, move-only struct with private fields that serves as a token denoting that funds have been destroyed without any equal and opposite accounting.

PositiveImbalance
Opaque, move-only struct with private fields that serves as a token denoting that funds have been created without any equal and opposite accounting.

ReservedBalance
The amount of the balance of a given account that is externally reserved; this can still get slashed, but gets slashed last of all.

TakeFees
Require the transactor pay for themselves and maybe include a tip to gain additional priority in the queue.

TotalIssuance
The total units issued in the system.

Vesting
Information regarding the vesting of a given account.

VestingSchedule
Struct to encode the vesting schedule of an individual account.

Enums

Call

RawEvent
Events for this module.

Constants

DEFAULT_CREATION_FEE

DEFAULT_EXISTENTIAL_DEPOSIT

DEFAULT_TRANSACTION_BASE_FEE

DEFAULT_TRANSACTION_BYTE_FEE

DEFAULT_TRANSFER_FEE

Traits

Instance
Tag a type as an instance of a module.

Subtrait

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait