Struct srml_balances::Module

pub struct Module<T: Trait<I>, I: Instance = DefaultInstance>(_);

Methods

impl<T: 'static + Trait<I>, I: Instance> Module<T, I>

pub fn total_issuance() -> T::Balance

The total units issued in the system.

pub fn vesting<K: Borrow<T::AccountId>>(
key: K
) -> Option<VestingSchedule<T::Balance, T::BlockNumber>>

Information regarding the vesting of a given account.

pub fn free_balance<K: Borrow<T::AccountId>>(key: K) -> T::Balance

The 'free' balance of a given account.

This is the only balance that matters in terms of most operations on tokens. It alone is used to determine the balance when in the contract execution environment. When this balance falls below the value of ExistentialDeposit, then the 'current account' is deleted: specifically FreeBalance. Further, the OnFreeBalanceZero callback is invoked, giving a chance to external modules to clean up data associated with the deleted account.

system::AccountNonce is also deleted if ReservedBalance is also zero (it also gets collapsed to zero if it ever becomes less than ExistentialDeposit.

pub fn reserved_balance<K: Borrow<T::AccountId>>(key: K) -> T::Balance

The amount of the balance of a given account that is externally reserved; this can still get slashed, but gets slashed last of all.

This balance is a 'reserve' balance that other subsystems use in order to set aside tokens that are still 'owned' by the account holder, but which are suspendable.

When this balance falls below the value of ExistentialDeposit, then this 'reserve account' is deleted: specifically, ReservedBalance.

system::AccountNonce is also deleted if FreeBalance is also zero (it also gets collapsed to zero if it ever becomes less than ExistentialDeposit.)

pub fn locks<K: Borrow<T::AccountId>>(
key: K
) -> Vec<BalanceLock<T::Balance, T::BlockNumber>>

Any liquidity locks on some account balances.

impl<T: Trait<I>, I: Instance> Module<T, I>

Can also be called using Call.

pub fn transfer(
origin: T::Origin,
dest: <T::Lookup as StaticLookup>::Source,
value: T::Balance
) -> Result

Transfer some liquid free balance to another account.

transfer will set the FreeBalance of the sender and receiver. It will decrease the total issuance of the system by the TransferFee. If the sender's account is below the existential deposit as a result of the transfer, the account will be reaped.

The dispatch origin for this call must be Signed by the transactor.

Dependent on arguments but not critical, given proper implementations for input config types. See related functions below.

It contains a limited number of reads and writes internally and no complex computation.

Related functions:

ensure_can_withdraw is always called internally but has a bounded complexity.

Transferring balances to accounts that did not exist before will cause T::OnNewAccount::on_new_account to be called.

Removing enough funds from an account will trigger T::DustRemoval::on_unbalanced and T::OnFreeBalanceZero::on_free_balance_zero.

impl<T: Trait<I>, I: Instance> Module<T, I>

pub fn vesting_balance(who: &T::AccountId) -> T::Balance

Get the amount that is currently being vested and cannot be transferred out of this account.