Struct srml_balances::NegativeImbalance

#[must_use]
pub struct NegativeImbalance<T: Subtrait<I>, I: Instance = DefaultInstance>(_);

Opaque, move-only struct with private fields that serves as a token denoting that funds have been destroyed without any equal and opposite accounting.

Methods

impl<T: Subtrait<I>, I: Instance> NegativeImbalance<T, I>

pub fn new(amount: T::Balance) -> Self

Create a new negative imbalance from a balance.